requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
attrs==24.2.0
av==12.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dnspython==2.7.0
entrypoints==0.4
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.10.0
numpy==1.26.4
opencv-python-headless
packaging==23.2
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==3.20.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pylibsrtp==0.10.0
Pympler==1.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
st-theme==1.2.3
streamlit==1.33.0
streamlit-js-eval==0.1.7
streamlit-navigation-bar==3.2.0
streamlit-webrtc==0.47.9
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
yt-dlp==2024.11.4
zipp==3.20.2